Move functions evaluating unit HP recovery to common
They will be needed to handle units with HP loss (helicopters) or simply units
that would benefit from healing before moving. Think horses with HP=1 in a city
with barracks: it will often be faster to wait one turn in the city until we
get full HP, then only start moving.

So the client needs to know about HP recovery.

One subtlety of the implementation is the killunhomed server setting, which was
only sent as a setting and not as an easier-to-user game_info field. So I added
it to the protocol under a capability so network compatibility is preserved.
