+++ common/networking/packets.def
@@ -593,6 +593,7 @@ PACKET_GAME_INFO = 16; sc, is-info
   SINT16 civil_war_bonus_unhappy;
   BOOL paradrop_to_transport;
   UINT16 granularity;
+  SINT16 killunhomed; add-cap(killunhomed-is-game-info)
 end
 
 PACKET_CALENDAR_INFO = 255; sc, is-info, handle-via-packet
