+++ common/unit_utils.h
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 1996-2021 Freeciv contributors
+// SPDX-FileCopyrightText: 1996-2021 Freeciv21 contributors
+// SPDX-FileCopyrightText: 2022 Louis Moureaux
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#pragma once
+
+struct unit;
+
+void unit_restore_hitpoints(unit *punit);
